+++ spec/requests/claims_spec.rb
@@ -58,11 +58,11 @@ RSpec.describe "Claims", type: :request do
           expect(response.body).to include "Continue"
         end
 
-        it "only returns results if the search term is more than three characters" do
-          get claim_path(StudentLoans.routing_name, "claim-school"), params: {school_search: "Pen"}
+        it "only returns results if the search term is more than two characters" do
+          get claim_path(StudentLoans.routing_name, "claim-school"), params: {school_search: "Pe"}
 
           expect(response.body).to include("There is a problem")
-          expect(response.body).to include("Enter the name of the school")
+          expect(response.body).to include("Enter the name or postcode of the school")
           expect(response.body).not_to include(schools(:penistone_grammar_school).name)
         end